Adjusted the instrumentation of exit
- changed the error message
- moved the files into `tstutil.h` - this prevents students from working
  around it and calling `orig_exit`
- fixed the other tests to use `exit_orig`
