+++ tests/testbonusheap.c
@@ -48,7 +48,7 @@ void run_bfi(const char *bf_file, const char *expected)
         destroy_ijvm();
         eprintf("Couldn't read the output of the ijvm for program %s", bf_file);
         remove("tmp_output");
-        exit(1);
+        orig_exit(1);
     }
     /* add NIL byte at the end of the buffer */
     buffer[bytes_read] = 0;
@@ -56,7 +56,7 @@ void run_bfi(const char *bf_file, const char *expected)
     if (memcmp(buffer, expected, strlen(expected)+1) != 0)
     {
         eprintf("Incorrect output, got %s expected %s", buffer, expected);
-        exit(1);
+        orig_exit(1);
     }
     else
         eprintf("Successfully executed %s", bf_file);
